ci: Add HDF5 to a windows build

+++ examples/hello/hdf5Reader/helloHDF5Reader_nompi.cpp
@@ -22,9 +22,9 @@ void ReadData(adios2::IO h5IO, adios2::Engine h5Reader, const std::string &name)
 
     if (var)
     {
-        int nDims = var.Shape().size();
+        size_t nDims = var.Shape().size();
         size_t totalSize = 1;
-        for (int i = 0; i < nDims; i++)
+        for (size_t i = 0; i < nDims; i++)
         {
             totalSize *= var.Shape()[i];
         }
